refactor(consttime): rename subtlex package to consttime
Descriptive name for the constant-time comparison helpers. Updated
consumers: password, sign. Error-prefix sentinels now use 'consttime:'.
